Preserve mask when copying wxBitmapRefData to avoid crashes
This crash was introduced in f878e6060a04d08b47f5c8c2bdc20f51bbc15548
and could happens inside wxBitmap::MSWBlendMaskWithAlpha() because the
code checked for a mask and alpha, but the mask could be lost after
AllocExclusive().

In practice, this happened e.g. when using 16-bit color as is the case
for Windows 7 Hyper-V VMs.

Closes https://github.com/wxWidgets/wxWidgets/pull/1695